Fix rrules for Symmetric and Diagonal constructors
Currently these definitions are extending `rrule(::typeof(T), x)` where
`T` is a type. However, `typeof(Diagonal) == UnionAll`, which means this
is not defining the method it looks like it might be defining. The only
reason this worked when originally implemented was that one of the
`rrule` definitions was for `rrule(UnionAll, Matrix)` and the other for
`rrule(UnionAll, Vector)`, so dispatch still worked.

This replaces these problematic `::typeof(T)`s with `::Type{<:T}`.