Extend Base.isapprox instead of defining cr_isapprox (#26)
Currently in the testing utilities we define an internal `cr_isapprox`
function that mimics `isapprox` except we only accept a subset of the
keyword arguments accepted by `isapprox`. Instead, we can just define
methods for `isapprox` on ChainRules-defined types for the purposes of
testing. That way we can more easily take arbitrary keyword arguments
and get Test's special-cased pretty printing for `@test a ≈ b`.