Add NotImplemented and NotImplementedRule
Sometimes you just don't implement something. Maybe it's hard, maybe
you're lazy like me, whatever. For such cases, there is the differential
`NotImplemented` and its associated rule `NotImplementedRule`.

This was born of not being able to figure out the rule for the scalar
multiple parameter in `BLAS.gemm`, but having implemented rules for the
matrix parameters.