Switch to `using LinearAlgebra.BLAS`
Currently some of the rules are calling `blascopy!` and `scal!` without
qualification, which fails, since we haven't done a full `using` for the
BLAS submodule. This adds a `using` and an `import` for `dot` on earlier
versions of Julia to avoid a name conflict between LinearAlgebra and
BLAS which has since been fixed.